Extract info items from components
